Split Do_Func into per-function kernels
Do_Func was one 3,310-line function holding two matches over fifty unrelated
operations -- a constant-folding branch and a row-loop branch -- with the two
sharing a dozen mutable locals.  Each arm is now a named function in
src/eval_y/func.rs, and Do_Func is 183 lines: gather the arguments, then
dispatch.

The gathered arguments become a FuncArgs struct rather than four parallel
arrays threaded through by name, so a kernel's signature says what it reads.

Also removes the goto emulation in the constant branch.  current_block_139
had exactly one live user: FuncOp::Min1 jumped to a shared "copy the first
argument" block that FuncOp::Max1 already had inlined.  Inlining it for Min1
too made all 42 assignments dead.

33 of the 84 kernels no longer need an unsafe block at all.

The point of the split is that a kernel is now a plain function over a node
and its arguments, so it can be tested without a FITS file, a parser or an
iterator.  The new tests do exactly that for the sixteen transcendental
kernels, ABS over both numeric sorts, MIN/MAX of two arguments, ARCTAN2,
ANGSEP and the null predicates -- including pinning ANGSEP's constant-folded
answer, which is where rsfitsio deliberately differs from CFITSIO 4.7.0.
