+++ src/eval_y.rs
@@ -20,6 +20,10 @@ use core::{cmp, ptr};
 use bytemuck::{cast_slice, cast_slice_mut};
 use libc::{calloc, free, malloc, memcpy, time, time_t};
 
+pub(crate) mod func;
+
+use func::FuncArgs;
+
 const APPROX: f64 = 1.0e-7;
 
 // Math function wrappers
@@ -3862,24 +3866,15 @@ pub(crate) fn angsep_calc(
 
 fn Do_Func(lParse: &mut ParseData, this_node_idx: usize) {
     unsafe {
-        let mut theParams: [usize; 10] = [0; 10];
-        let mut vector: [c_int; 10] = [0; 10];
+        let mut args = FuncArgs::default();
         let mut allConst: c_int = 0;
-        let mut pVals: [lval; 10] = [lval {
-            nelem: 0,
-            naxis: 0,
-            naxes: [0; 5],
-            undef: core::ptr::null_mut::<c_char>(),
-            data: NodeValue::Double(0.0),
-        }; 10];
-        let mut pNull: [c_char; 10] = [0; 10];
-        let mut ival: c_long = 0;
-        let mut dval: c_double = 0.0;
+        let ival: c_long = 0;
+        let dval: c_double = 0.0;
         let mut i: c_int = 0;
-        let mut valInit: c_int = 0;
+        let valInit: c_int = 0;
         let mut row: c_long = 0;
         let mut elem: c_long = 0;
-        let mut nelem: c_long = 0;
+        let nelem: c_long = 0;
         /* Do_Func is only installed for function nodes, so `operation` is
         always a FuncOp here. */
         let Operation::Func(op) = (lParse.Nodes[this_node_idx]).operation else {
@@ -3895,32 +3890,34 @@ fn Do_Func(lParse: &mut ParseData, this_node_idx: usize) {
             if fresh52 == 0 {
                 break;
             }
-            theParams[i as usize] = (lParse.Nodes[this_node_idx]).SubNodes[i as usize] as usize;
-            vector[i as usize] =
-                c_int::from((lParse.Nodes[theParams[i as usize]]).operation != Operation::Const);
-            if vector[i as usize] != 0 {
+            args.params[i as usize] = (lParse.Nodes[this_node_idx]).SubNodes[i as usize] as usize;
+            args.vector[i as usize] =
+                c_int::from((lParse.Nodes[args.params[i as usize]]).operation != Operation::Const);
+            if args.vector[i as usize] != 0 {
                 allConst = 0;
-                vector[i as usize] = (lParse.Nodes[theParams[i as usize]]).value.nelem as c_int;
+                args.vector[i as usize] =
+                    (lParse.Nodes[args.params[i as usize]]).value.nelem as c_int;
             } else {
-                if (lParse.Nodes[theParams[i as usize]]).ntype == ValueSort::Double {
-                    pVals[i as usize].data =
-                        NodeValue::Double((lParse.Nodes[theParams[i as usize]]).value.data.dbl());
-                } else if (lParse.Nodes[theParams[i as usize]]).ntype == ValueSort::Long {
-                    pVals[i as usize].data =
-                        NodeValue::Long((lParse.Nodes[theParams[i as usize]]).value.data.lng());
-                } else if (lParse.Nodes[theParams[i as usize]]).ntype == ValueSort::Boolean {
-                    pVals[i as usize].data =
-                        NodeValue::Logical((lParse.Nodes[theParams[i as usize]]).value.data.log());
+                if (lParse.Nodes[args.params[i as usize]]).ntype == ValueSort::Double {
+                    args.vals[i as usize].data =
+                        NodeValue::Double((lParse.Nodes[args.params[i as usize]]).value.data.dbl());
+                } else if (lParse.Nodes[args.params[i as usize]]).ntype == ValueSort::Long {
+                    args.vals[i as usize].data =
+                        NodeValue::Long((lParse.Nodes[args.params[i as usize]]).value.data.lng());
+                } else if (lParse.Nodes[args.params[i as usize]]).ntype == ValueSort::Boolean {
+                    args.vals[i as usize].data = NodeValue::Logical(
+                        (lParse.Nodes[args.params[i as usize]]).value.data.log(),
+                    );
                 } else {
                     strcpy(
-                        pVals[i as usize].data.text_mut_ptr(),
-                        (lParse.Nodes[theParams[i as usize]])
+                        args.vals[i as usize].data.text_mut_ptr(),
+                        (lParse.Nodes[args.params[i as usize]])
                             .value
                             .data
                             .text_mut_ptr(),
                     );
                 }
-                pNull[i as usize] = 0;
+                args.nulls[i as usize] = 0;
             }
         }
         if (lParse.Nodes[this_node_idx]).nSubNodes == 0 {
@@ -3936,421 +3933,47 @@ fn Do_Func(lParse: &mut ParseData, this_node_idx: usize) {
             allConst = 0;
         }
         if allConst != 0 {
-            let current_block_139: u64;
             match op {
-                FuncOp::Sum => {
-                    if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Boolean {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(c_long::from(
-                            if c_int::from(pVals[0].data.log()) != 0 {
-                                1
-                            } else {
-                                0
-                            },
-                        ));
-                    } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(pVals[0].data.lng());
-                    } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Bits {
-                        strcpy(
-                            (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr(),
-                            pVals[0].data.text_mut_ptr(),
-                        );
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Average => {
-                    if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.lng() as c_double);
-                    } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Stddev => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(0.0);
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Median => {
-                    if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Boolean {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(c_long::from(
-                            if c_int::from(pVals[0].data.log()) != 0 {
-                                1
-                            } else {
-                                0
-                            },
-                        ));
-                    } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(pVals[0].data.lng());
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::PoiRnd => {
-                    if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(c_long::from(
-                            simplerng_getpoisson(pVals[0].data.dbl()),
-                        ));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(c_long::from(
-                            simplerng_getpoisson(pVals[0].data.lng() as c_double),
-                        ));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Abs => {
-                    if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                        dval = pVals[0].data.dbl();
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(if dval > 0.0 { dval } else { -dval });
-                    } else {
-                        ival = pVals[0].data.lng();
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(if ival > 0 { ival } else { -ival });
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::NonNull => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(1);
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::IsNull => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(0);
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::DefNull => {
-                    if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Boolean {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Logical(pVals[0].data.log());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(pVals[0].data.lng());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::String {
-                        strcpy(
-                            (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr(),
-                            pVals[0].data.text_mut_ptr(),
-                        );
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::SetNull => {
-                    if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(pVals[0].data.lng());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Sin => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(sin(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Cos => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(cos(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Tan => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(tan(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Asin => {
-                    dval = pVals[0].data.dbl();
-                    if dval < -1.0 || dval > 1.0 {
-                        fits_parser_yyerror(lParse, cs!(c"Out of range argument to arcsin"));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(asin(dval));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Acos => {
-                    dval = pVals[0].data.dbl();
-                    if dval < -1.0 || dval > 1.0 {
-                        fits_parser_yyerror(lParse, cs!(c"Out of range argument to arccos"));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(acos(dval));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Atan => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(atan(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Sinh => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(sinh(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Cosh => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(cosh(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Tanh => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(tanh(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Exp => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(exp(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Log => {
-                    dval = pVals[0].data.dbl();
-                    if dval <= 0.0 {
-                        fits_parser_yyerror(lParse, cs!(c"Out of range argument to log"));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(log(dval));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Log10 => {
-                    dval = pVals[0].data.dbl();
-                    if dval <= 0.0 {
-                        fits_parser_yyerror(lParse, cs!(c"Out of range argument to log10"));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(log10(dval));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Sqrt => {
-                    dval = pVals[0].data.dbl();
-                    if dval < 0.0 {
-                        fits_parser_yyerror(lParse, cs!(c"Out of range argument to sqrt"));
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(sqrt(dval));
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Ceil => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(ceil(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Floor => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(floor(pVals[0].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Round => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(floor(pVals[0].data.dbl() + 0.5));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Atan2 => {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(atan2(pVals[0].data.dbl(), pVals[1].data.dbl()));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::AngSep => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Double(angsep_calc(
-                        pVals[0].data.dbl(),
-                        pVals[1].data.dbl(),
-                        pVals[2].data.dbl(),
-                        pVals[3].data.dbl(),
-                    ));
-                    /* DEVIATION from CFITSIO 4.7.0: "case angsep_fct:" there
-                    is missing its "break;" and falls through into
-                    "case min1_fct:", which overwrites the separation just
-                    computed with pVals[0] - so a constant-folded ANGSEP
-                    returns its first argument.  Fix submitted upstream; the
-                    non-constant path in Do_Func has always been correct. */
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Min1 => {
-                    current_block_139 = 15934000668868306918;
-                }
-                FuncOp::Min2 => {
-                    if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(if pVals[0].data.dbl() < pVals[1].data.dbl() {
-                                pVals[0].data.dbl()
-                            } else {
-                                pVals[1].data.dbl()
-                            });
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(if pVals[0].data.lng() < pVals[1].data.lng() {
-                                pVals[0].data.lng()
-                            } else {
-                                pVals[1].data.lng()
-                            });
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Max1 => {
-                    if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(pVals[0].data.dbl());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(pVals[0].data.lng());
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Bits {
-                        strcpy(
-                            (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr(),
-                            pVals[0].data.text_mut_ptr(),
-                        );
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Max2 => {
-                    if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Double(if pVals[0].data.dbl() > pVals[1].data.dbl() {
-                                pVals[0].data.dbl()
-                            } else {
-                                pVals[1].data.dbl()
-                            });
-                    } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                        (lParse.Nodes[this_node_idx]).value.data =
-                            NodeValue::Long(if pVals[0].data.lng() > pVals[1].data.lng() {
-                                pVals[0].data.lng()
-                            } else {
-                                pVals[1].data.lng()
-                            });
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Near => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(bnear(
-                        pVals[0].data.dbl(),
-                        pVals[1].data.dbl(),
-                        pVals[2].data.dbl(),
-                    ));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Circle => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(circle(
-                        pVals[0].data.dbl(),
-                        pVals[1].data.dbl(),
-                        pVals[2].data.dbl(),
-                        pVals[3].data.dbl(),
-                        pVals[4].data.dbl(),
-                    ));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Box => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(saobox(
-                        pVals[0].data.dbl(),
-                        pVals[1].data.dbl(),
-                        pVals[2].data.dbl(),
-                        pVals[3].data.dbl(),
-                        pVals[4].data.dbl(),
-                        pVals[5].data.dbl(),
-                        pVals[6].data.dbl(),
-                    ));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::Ellipse => {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(ellipse(
-                        pVals[0].data.dbl(),
-                        pVals[1].data.dbl(),
-                        pVals[2].data.dbl(),
-                        pVals[3].data.dbl(),
-                        pVals[4].data.dbl(),
-                        pVals[5].data.dbl(),
-                        pVals[6].data.dbl(),
-                    ));
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::IfThenElse => {
-                    match (lParse.Nodes[this_node_idx]).ntype {
-                        ValueSort::Boolean => {
-                            (lParse.Nodes[this_node_idx]).value.data = NodeValue::Logical(
-                                (if c_int::from(pVals[2].data.log()) != 0 {
-                                    c_int::from(pVals[0].data.log())
-                                } else {
-                                    c_int::from(pVals[1].data.log())
-                                }) as c_char,
-                            );
-                        }
-                        ValueSort::Long => {
-                            (lParse.Nodes[this_node_idx]).value.data =
-                                NodeValue::Long(if c_int::from(pVals[2].data.log()) != 0 {
-                                    pVals[0].data.lng()
-                                } else {
-                                    pVals[1].data.lng()
-                                });
-                        }
-                        ValueSort::Double => {
-                            (lParse.Nodes[this_node_idx]).value.data =
-                                NodeValue::Double(if c_int::from(pVals[2].data.log()) != 0 {
-                                    pVals[0].data.dbl()
-                                } else {
-                                    pVals[1].data.dbl()
-                                });
-                        }
-                        ValueSort::String => {
-                            strcpy(
-                                (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr(),
-                                if c_int::from(pVals[2].data.log()) != 0 {
-                                    pVals[0].data.text_mut_ptr()
-                                } else {
-                                    pVals[1].data.text_mut_ptr()
-                                },
-                            );
-                        }
-                        _ => {}
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::StrMid => {
-                    let dest_str = (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr();
-                    let dest_len = (lParse.Nodes[this_node_idx]).value.nelem as c_int;
-                    cstrmid(
-                        lParse,
-                        dest_str,
-                        dest_len,
-                        pVals[0].data.text_mut_ptr(),
-                        pVals[0].nelem as c_int,
-                        pVals[1].data.lng() as c_int,
-                    );
-                    current_block_139 = 7627602990488000394;
-                }
-                FuncOp::StrPos => {
-                    let res: *mut c_char =
-                        strstr(pVals[0].data.text_mut_ptr(), pVals[1].data.text_mut_ptr());
-                    if res.is_null() {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(0);
-                    } else {
-                        (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(
-                            res.offset_from(pVals[0].data.text_mut_ptr()) as c_long + 1,
-                        );
-                    }
-                    current_block_139 = 7627602990488000394;
-                }
-                _ => {
-                    current_block_139 = 7627602990488000394;
-                }
-            }
-            if current_block_139 == 15934000668868306918 {
-                if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                    (lParse.Nodes[this_node_idx]).value.data =
-                        NodeValue::Double(pVals[0].data.dbl());
-                } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                    (lParse.Nodes[this_node_idx]).value.data = NodeValue::Long(pVals[0].data.lng());
-                } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Bits {
-                    strcpy(
-                        (lParse.Nodes[this_node_idx]).value.data.text_mut_ptr(),
-                        pVals[0].data.text_mut_ptr(),
-                    );
-                }
+                FuncOp::Sum => func::sum_const(lParse, this_node_idx, &mut args),
+                FuncOp::Average => func::average_const(lParse, this_node_idx, &mut args),
+                FuncOp::Stddev => func::stddev_const(lParse, this_node_idx, &mut args),
+                FuncOp::Median => func::median_const(lParse, this_node_idx, &mut args),
+                FuncOp::PoiRnd => func::poi_rnd_const(lParse, this_node_idx, &mut args),
+                FuncOp::Abs => func::abs_const(lParse, this_node_idx, &mut args),
+                FuncOp::NonNull => func::non_null_const(lParse, this_node_idx, &mut args),
+                FuncOp::IsNull => func::is_null_const(lParse, this_node_idx, &mut args),
+                FuncOp::DefNull => func::def_null_const(lParse, this_node_idx, &mut args),
+                FuncOp::SetNull => func::set_null_const(lParse, this_node_idx, &mut args),
+                FuncOp::Sin => func::sin_const(lParse, this_node_idx, &mut args),
+                FuncOp::Cos => func::cos_const(lParse, this_node_idx, &mut args),
+                FuncOp::Tan => func::tan_const(lParse, this_node_idx, &mut args),
+                FuncOp::Asin => func::asin_const(lParse, this_node_idx, &mut args),
+                FuncOp::Acos => func::acos_const(lParse, this_node_idx, &mut args),
+                FuncOp::Atan => func::atan_const(lParse, this_node_idx, &mut args),
+                FuncOp::Sinh => func::sinh_const(lParse, this_node_idx, &mut args),
+                FuncOp::Cosh => func::cosh_const(lParse, this_node_idx, &mut args),
+                FuncOp::Tanh => func::tanh_const(lParse, this_node_idx, &mut args),
+                FuncOp::Exp => func::exp_const(lParse, this_node_idx, &mut args),
+                FuncOp::Log => func::log_const(lParse, this_node_idx, &mut args),
+                FuncOp::Log10 => func::log10_const(lParse, this_node_idx, &mut args),
+                FuncOp::Sqrt => func::sqrt_const(lParse, this_node_idx, &mut args),
+                FuncOp::Ceil => func::ceil_const(lParse, this_node_idx, &mut args),
+                FuncOp::Floor => func::floor_const(lParse, this_node_idx, &mut args),
+                FuncOp::Round => func::round_const(lParse, this_node_idx, &mut args),
+                FuncOp::Atan2 => func::atan2_const(lParse, this_node_idx, &mut args),
+                FuncOp::AngSep => func::ang_sep_const(lParse, this_node_idx, &mut args),
+                FuncOp::Min1 => func::min1_const(lParse, this_node_idx, &mut args),
+                FuncOp::Min2 => func::min2_const(lParse, this_node_idx, &mut args),
+                FuncOp::Max1 => func::max1_const(lParse, this_node_idx, &mut args),
+                FuncOp::Max2 => func::max2_const(lParse, this_node_idx, &mut args),
+                FuncOp::Near => func::near_const(lParse, this_node_idx, &mut args),
+                FuncOp::Circle => func::circle_const(lParse, this_node_idx, &mut args),
+                FuncOp::Box => func::box_const(lParse, this_node_idx, &mut args),
+                FuncOp::Ellipse => func::ellipse_const(lParse, this_node_idx, &mut args),
+                FuncOp::IfThenElse => func::if_then_else_const(lParse, this_node_idx, &mut args),
+                FuncOp::StrMid => func::str_mid_const(lParse, this_node_idx, &mut args),
+                FuncOp::StrPos => func::str_pos_const(lParse, this_node_idx, &mut args),
+                _ => {}
             }
             (lParse.Nodes[this_node_idx]).operation = Operation::Const;
         } else {
@@ -4359,2796 +3982,51 @@ fn Do_Func(lParse: &mut ParseData, this_node_idx: usize) {
             elem = row * (lParse.Nodes[this_node_idx]).value.nelem;
             if lParse.status == 0 {
                 match op {
-                    FuncOp::Row => loop {
-                        let fresh53 = row;
-                        row -= 1;
-                        if fresh53 == 0 {
-                            break;
-                        }
-                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                            .offset(row as isize) = lParse.firstRow + row;
-                        *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) = 0;
-                    },
-                    FuncOp::Null => {
-                        if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh54 = row;
-                                row -= 1;
-                                if fresh54 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::String {
-                            loop {
-                                let fresh55 = row;
-                                row -= 1;
-                                if fresh55 == 0 {
-                                    break;
-                                }
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(0) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                            }
-                        }
-                    }
-                    FuncOp::AxisElem => {
-                        let mut ielem: c_long = 0;
-                        let mut iaxis: [c_long; 5] = [1, 1, 1, 1, 1];
-                        let ipos: c_long = pVals[1].data.lng() - 1;
-                        let naxis: c_int = (lParse.Nodes[this_node_idx]).value.naxis;
-                        let mut j: c_int = 0;
-                        if ipos < 0 || ipos >= 5 as c_long {
-                            fits_parser_yyerror(
-                                lParse,
-                                cs!(c"AXISELEM(V,n) n value exceeded maximum dimension"),
-                            );
-                            free((lParse.Nodes[this_node_idx]).value.data.raw());
-                        } else {
-                            ielem = 0;
-                            while ielem < elem {
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(ielem as isize) = iaxis[ipos as usize];
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(ielem as isize) = 0;
-                                iaxis[0] += 1;
-                                j = 0;
-                                while j < naxis {
-                                    if iaxis[j as usize]
-                                        <= (lParse.Nodes[this_node_idx]).value.naxes[j as usize]
-                                    {
-                                        break;
-                                    }
-                                    iaxis[j as usize] = 1;
-                                    if j < naxis - 1 {
-                                        iaxis[(j + 1) as usize] += 1;
-                                    }
-                                    j += 1;
-                                }
-                                ielem += 1;
-                            }
-                        }
-                    }
-                    FuncOp::ElemNum => {
-                        let mut ielem_0: c_long = 0;
-                        let mut elemnum: c_long = 1;
-                        let j_0: c_int = 0;
-                        ielem_0 = 0;
-                        while ielem_0 < elem {
-                            *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                .offset(ielem_0 as isize) = elemnum;
-                            *((lParse.Nodes[this_node_idx]).value.undef).offset(ielem_0 as isize) =
-                                0;
-                            elemnum += 1;
-                            if elemnum > (lParse.Nodes[this_node_idx]).value.nelem {
-                                elemnum = 1;
-                            }
-                            ielem_0 += 1;
-                        }
-                    }
-                    FuncOp::Rnd => loop {
-                        let fresh56 = elem;
-                        elem -= 1;
-                        if fresh56 == 0 {
-                            break;
-                        }
-                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                            .offset(elem as isize) = simplerng_getuniform();
-                        *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize) = 0;
-                    },
-                    FuncOp::GasRnd => loop {
-                        let fresh57 = elem;
-                        elem -= 1;
-                        if fresh57 == 0 {
-                            break;
-                        }
-                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                            .offset(elem as isize) = simplerng_getnorm();
-                        *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize) = 0;
-                    },
-                    FuncOp::PoiRnd => {
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                            if (lParse.Nodes[theParams[0]]).operation == Operation::Const {
-                                loop {
-                                    let fresh58 = elem;
-                                    elem -= 1;
-                                    if fresh58 == 0 {
-                                        break;
-                                    }
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) =
-                                        if pVals[0].data.dbl() < 0.0 { 1 } else { 0 };
-                                    if *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) =
-                                            c_long::from(simplerng_getpoisson(pVals[0].data.dbl()));
-                                    }
-                                }
-                            } else {
-                                loop {
-                                    let fresh59 = elem;
-                                    elem -= 1;
-                                    if fresh59 == 0 {
-                                        break;
-                                    }
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) =
-                                        *((lParse.Nodes[theParams[0]]).value.undef)
-                                            .offset(elem as isize);
-                                    if *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                        .offset(elem as isize)
-                                        < 0.0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 1;
-                                    }
-                                    if *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) =
-                                            c_long::from(simplerng_getpoisson(
-                                                *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(elem as isize),
-                                            ));
-                                    }
-                                }
-                            }
-                        } else if (lParse.Nodes[theParams[0]]).operation == Operation::Const {
-                            loop {
-                                let fresh60 = elem;
-                                elem -= 1;
-                                if fresh60 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    if pVals[0].data.lng() < 0 { 1 } else { 0 };
-                                if *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize)
-                                    == 0
-                                {
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(elem as isize) = c_long::from(
-                                        simplerng_getpoisson(pVals[0].data.lng() as c_double),
-                                    );
-                                }
-                            }
-                        } else {
-                            loop {
-                                let fresh61 = elem;
-                                elem -= 1;
-                                if fresh61 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize);
-                                if *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                    .offset(elem as isize)
-                                    < 0
-                                {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = 1;
-                                }
-                                if *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize)
-                                    == 0
-                                {
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(elem as isize) =
-                                        c_long::from(simplerng_getpoisson(
-                                            *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                                .offset(elem as isize)
-                                                as c_double,
-                                        ));
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Sum => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Boolean {
-                            loop {
-                                let fresh62 = row;
-                                row -= 1;
-                                if fresh62 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh63 = nelem;
-                                    nelem -= 1;
-                                    if fresh63 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(row as isize) += c_long::from(
-                                            if c_int::from(
-                                                *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .log_buf())
-                                                .offset(elem as isize),
-                                            ) != 0
-                                            {
-                                                1
-                                            } else {
-                                                0
-                                            },
-                                        );
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                            }
-                        } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh64 = row;
-                                row -= 1;
-                                if fresh64 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh65 = nelem;
-                                    nelem -= 1;
-                                    if fresh65 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(row as isize) +=
-                                            *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                                .offset(elem as isize);
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                            }
-                        } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh66 = row;
-                                row -= 1;
-                                if fresh66 == 0 {
-                                    break;
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(row as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh67 = nelem;
-                                    nelem -= 1;
-                                    if fresh67 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(row as isize) +=
-                                            *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                                .offset(elem as isize);
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                            }
-                        } else {
-                            nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                            loop {
-                                let fresh68 = row;
-                                row -= 1;
-                                if fresh68 == 0 {
-                                    break;
-                                }
-                                let mut sptr1: *mut c_char =
-                                    *((lParse.Nodes[theParams[0]]).value.data.str_buf())
-                                        .offset(row as isize);
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    0;
-                                while *sptr1 != 0 {
-                                    if c_int::from(*sptr1) == '1' as i32 {
-                                        let fresh69 = &mut *((lParse.Nodes[this_node_idx])
-                                            .value
-                                            .data
-                                            .lng_buf())
-                                        .offset(row as isize);
-                                        *fresh69 += 1;
-                                    }
-                                    sptr1 = sptr1.offset(1);
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Average => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh70 = row;
-                                row -= 1;
-                                if fresh70 == 0 {
-                                    break;
-                                }
-                                let mut count: c_int = 0;
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(row as isize) = 0.0;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh71 = nelem;
-                                    nelem -= 1;
-                                    if fresh71 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if c_int::from(
-                                        *((lParse.Nodes[theParams[0]]).value.undef)
-                                            .offset(elem as isize),
-                                    ) == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(row as isize) +=
-                                            *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                                .offset(elem as isize)
-                                                as c_double;
-                                        count += 1;
-                                    }
-                                }
-                                if count == 0 {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 1;
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) /= c_double::from(count);
-                                }
-                            }
-                        } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh72 = row;
-                                row -= 1;
-                                if fresh72 == 0 {
-                                    break;
-                                }
-                                let mut count_0: c_int = 0;
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(row as isize) = 0.0;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh73 = nelem;
-                                    nelem -= 1;
-                                    if fresh73 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if c_int::from(
-                                        *((lParse.Nodes[theParams[0]]).value.undef)
-                                            .offset(elem as isize),
-                                    ) == 0
-                                    {
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(row as isize) +=
-                                            *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                                .offset(elem as isize);
-                                        count_0 += 1;
-                                    }
-                                }
-                                if count_0 == 0 {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 1;
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) /= c_double::from(count_0);
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Stddev => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh74 = row;
-                                row -= 1;
-                                if fresh74 == 0 {
-                                    break;
-                                }
-                                let mut count_1: c_int = 0;
-                                let mut sum: c_double = 0.0;
-                                let mut sum2: c_double = 0.0;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh75 = nelem;
-                                    nelem -= 1;
-                                    if fresh75 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if c_int::from(
-                                        *((lParse.Nodes[theParams[0]]).value.undef)
-                                            .offset(elem as isize),
-                                    ) == 0
-                                    {
-                                        sum += *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                            .offset(elem as isize)
-                                            as c_double;
-                                        count_1 += 1;
-                                    }
-                                }
-                                if count_1 > 1 {
-                                    sum /= c_double::from(count_1);
-                                    nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                    elem += nelem;
-                                    loop {
-                                        let fresh76 = nelem;
-                                        nelem -= 1;
-                                        if fresh76 == 0 {
-                                            break;
-                                        }
-                                        elem -= 1;
-
-                                        if c_int::from(
-                                            *((lParse.Nodes[theParams[0]]).value.undef)
-                                                .offset(elem as isize),
-                                        ) == 0
-                                        {
-                                            let dx: c_double = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(elem as isize)
-                                                as c_double
-                                                - sum;
-                                            sum2 += dx * dx;
-                                        }
-                                    }
-                                    sum2 /= c_double::from(count_1) - c_double::from(1);
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) = sqrt(sum2);
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) = 0.0;
-                                }
-                            }
-                        } else if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh77 = row;
-                                row -= 1;
-                                if fresh77 == 0 {
-                                    break;
-                                }
-                                let mut count_2: c_int = 0;
-                                let mut sum_0: c_double = 0.0;
-                                let mut sum2_0: c_double = 0.0;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh78 = nelem;
-                                    nelem -= 1;
-                                    if fresh78 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-
-                                    if c_int::from(
-                                        *((lParse.Nodes[theParams[0]]).value.undef)
-                                            .offset(elem as isize),
-                                    ) == 0
-                                    {
-                                        sum_0 +=
-                                            *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                                .offset(elem as isize);
-                                        count_2 += 1;
-                                    }
-                                }
-                                if count_2 > 1 {
-                                    sum_0 /= c_double::from(count_2);
-                                    nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                    elem += nelem;
-                                    loop {
-                                        let fresh79 = nelem;
-                                        nelem -= 1;
-                                        if fresh79 == 0 {
-                                            break;
-                                        }
-                                        elem -= 1;
-
-                                        if c_int::from(
-                                            *((lParse.Nodes[theParams[0]]).value.undef)
-                                                .offset(elem as isize),
-                                        ) == 0
-                                        {
-                                            let dx_0: c_double = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(elem as isize)
-                                                - sum_0;
-                                            sum2_0 += dx_0 * dx_0;
-                                        }
-                                    }
-                                    sum2_0 /= c_double::from(count_2) - c_double::from(1);
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) = sqrt(sum2_0);
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(row as isize) = 0.0;
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Median => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Long {
-                            let mut dptr: *mut c_long =
-                                (lParse.Nodes[theParams[0]]).value.data.lng_buf();
-                            let mut uptr: *mut c_char = (lParse.Nodes[theParams[0]]).value.undef;
-                            let mptr: *mut c_long = malloc(
-                                (::core::mem::size_of::<c_long>() as c_ulong)
-                                    .wrapping_mul(nelem as c_ulong)
-                                    .try_into()
-                                    .unwrap(),
-                            )
-                            .cast::<c_long>();
-                            let mut irow: c_int = 0;
-                            if mptr.is_null() {
-                                fits_parser_yyerror(
-                                    lParse,
-                                    cs!(c"Could not allocate temporary memory in median function"),
-                                );
-                                free((lParse.Nodes[this_node_idx]).value.data.raw());
-                            } else {
-                                irow = 0;
-                                while c_long::from(irow) < row {
-                                    let mut p: *mut c_long = mptr;
-                                    let mut nelem1: c_int = nelem as c_int;
-                                    loop {
-                                        let fresh80 = nelem1;
-                                        nelem1 -= 1;
-                                        if fresh80 == 0 {
-                                            break;
-                                        }
-                                        if c_int::from(*uptr) == 0 {
-                                            let fresh81 = p;
-                                            p = p.offset(1);
-                                            *fresh81 = *dptr;
-                                        }
-                                        dptr = dptr.offset(1);
-                                        uptr = uptr.offset(1);
-                                    }
-                                    nelem1 = p.offset_from(mptr) as c_long as c_int;
-                                    if nelem1 > 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(irow as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(irow as isize) = qselect_median(
-                                            slice::from_raw_parts_mut(mptr, nelem1 as usize),
-                                        );
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(irow as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(irow as isize) = 0;
-                                    }
-                                    irow += 1;
-                                }
-                                free(mptr.cast::<c_void>());
-                            }
-                        } else {
-                            let mut dptr_0: *mut c_double =
-                                (lParse.Nodes[theParams[0]]).value.data.dbl_buf();
-                            let mut uptr_0: *mut c_char = (lParse.Nodes[theParams[0]]).value.undef;
-                            let mptr_0: *mut c_double = malloc(
-                                (::core::mem::size_of::<c_double>() as c_ulong)
-                                    .wrapping_mul(nelem as c_ulong)
-                                    .try_into()
-                                    .unwrap(),
-                            )
-                            .cast::<c_double>();
-                            let mut irow_0: c_int = 0;
-                            if mptr_0.is_null() {
-                                fits_parser_yyerror(
-                                    lParse,
-                                    cs!(c"Could not allocate temporary memory in median function"),
-                                );
-                                free((lParse.Nodes[this_node_idx]).value.data.raw());
-                            } else {
-                                irow_0 = 0;
-                                while c_long::from(irow_0) < row {
-                                    let mut p_0: *mut c_double = mptr_0;
-                                    let mut nelem1_0: c_int = nelem as c_int;
-                                    loop {
-                                        let fresh82 = nelem1_0;
-                                        nelem1_0 -= 1;
-                                        if fresh82 == 0 {
-                                            break;
-                                        }
-                                        if c_int::from(*uptr_0) == 0 {
-                                            let fresh83 = p_0;
-                                            p_0 = p_0.offset(1);
-                                            *fresh83 = *dptr_0;
-                                        }
-                                        dptr_0 = dptr_0.offset(1);
-                                        uptr_0 = uptr_0.offset(1);
-                                    }
-                                    nelem1_0 = p_0.offset_from(mptr_0) as c_long as c_int;
-                                    if nelem1_0 > 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(irow_0 as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(irow_0 as isize) = qselect_median(
-                                            slice::from_raw_parts_mut(mptr_0, nelem1_0 as usize),
-                                        );
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(irow_0 as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(irow_0 as isize) = 0.0;
-                                    }
-                                    irow_0 += 1;
-                                }
-                                free(mptr_0.cast::<c_void>());
-                            }
-                        }
-                    }
-                    FuncOp::Abs => {
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh84 = elem;
-                                elem -= 1;
-                                if fresh84 == 0 {
-                                    break;
-                                }
-                                dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize);
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = if dval > 0.0 { dval } else { -dval };
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize);
-                            }
-                        } else {
-                            loop {
-                                let fresh85 = elem;
-                                elem -= 1;
-                                if fresh85 == 0 {
-                                    break;
-                                }
-                                ival = *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                    .offset(elem as isize);
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(elem as isize) = if ival > 0 { ival } else { -ival };
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize);
-                            }
-                        }
-                    }
-                    FuncOp::NonNull => {
-                        nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::String {
-                            nelem = 1;
-                        }
-                        elem = row * nelem;
-                        loop {
-                            let fresh86 = row;
-                            row -= 1;
-                            if fresh86 == 0 {
-                                break;
-                            }
-                            let mut nelem1_1: c_int = nelem as c_int;
-                            *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) = 0;
-                            *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                .offset(row as isize) = 0;
-                            loop {
-                                let fresh87 = nelem1_1;
-                                nelem1_1 -= 1;
-                                if fresh87 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-
-                                if c_int::from(
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize),
-                                ) == 0
-                                {
-                                    let fresh88 =
-                                        &mut *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(row as isize);
-                                    *fresh88 += 1;
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::IsNull => {
-                        if (lParse.Nodes[theParams[0]]).ntype == ValueSort::String {
-                            elem = row;
-                        }
-                        loop {
-                            let fresh89 = elem;
-                            elem -= 1;
-                            if fresh89 == 0 {
-                                break;
-                            }
-                            *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                .offset(elem as isize) =
-                                *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                            *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize) = 0;
-                        }
-                    }
-                    FuncOp::DefNull => match (lParse.Nodes[this_node_idx]).ntype {
-                        ValueSort::Boolean => loop {
-                            let fresh90 = row;
-                            row -= 1;
-                            if fresh90 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh91 = nelem;
-                                nelem -= 1;
-                                if fresh91 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-
-                                i = 2;
-                                loop {
-                                    let fresh92 = i;
-                                    i -= 1;
-                                    if fresh92 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                        pVals[i as usize].data = NodeValue::Logical(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .log_buf())
-                                            .offset(elem as isize),
-                                        );
-                                    } else if vector[i as usize] != 0 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                        pVals[i as usize].data = NodeValue::Logical(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .log_buf())
-                                            .offset(row as isize),
-                                        );
-                                    }
-                                }
-                                if pNull[0] != 0 {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = pNull[1];
-                                    *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                        .offset(elem as isize) = pVals[1].data.log();
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                        .offset(elem as isize) = pVals[0].data.log();
-                                }
-                            }
-                        },
-                        ValueSort::Long => loop {
-                            let fresh93 = row;
-                            row -= 1;
-                            if fresh93 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh94 = nelem;
-                                nelem -= 1;
-                                if fresh94 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-
-                                i = 2;
-                                loop {
-                                    let fresh95 = i;
-                                    i -= 1;
-                                    if fresh95 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                        pVals[i as usize].data = NodeValue::Long(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(elem as isize),
-                                        );
-                                    } else if vector[i as usize] != 0 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                        pVals[i as usize].data = NodeValue::Long(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(row as isize),
-                                        );
-                                    }
-                                }
-                                if pNull[0] != 0 {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = pNull[1];
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(elem as isize) = pVals[1].data.lng();
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(elem as isize) = pVals[0].data.lng();
-                                }
-                            }
-                        },
-                        ValueSort::Double => loop {
-                            let fresh96 = row;
-                            row -= 1;
-                            if fresh96 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh97 = nelem;
-                                nelem -= 1;
-                                if fresh97 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-
-                                i = 2;
-                                loop {
-                                    let fresh98 = i;
-                                    i -= 1;
-                                    if fresh98 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                        pVals[i as usize].data = NodeValue::Double(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(elem as isize),
-                                        );
-                                    } else if vector[i as usize] != 0 {
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                        pVals[i as usize].data = NodeValue::Double(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(row as isize),
-                                        );
-                                    }
-                                }
-                                if pNull[0] != 0 {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = pNull[1];
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(elem as isize) = pVals[1].data.dbl();
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(elem as isize) = 0;
-                                    *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                        .offset(elem as isize) = pVals[0].data.dbl();
-                                }
-                            }
-                        },
-                        ValueSort::String => loop {
-                            let fresh99 = row;
-                            row -= 1;
-                            if fresh99 == 0 {
-                                break;
-                            }
-                            i = 2;
-                            loop {
-                                let fresh100 = i;
-                                i -= 1;
-                                if fresh100 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] != 0 {
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                    strcpy(
-                                        pVals[i as usize].data.text_mut_ptr(),
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .str_buf())
-                                        .offset(row as isize),
-                                    );
-                                }
-                            }
-                            if pNull[0] != 0 {
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    pNull[1];
-                                strcpy(
-                                    *((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                        .offset(row as isize),
-                                    pVals[1].data.text_mut_ptr(),
-                                );
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 0;
-                                strcpy(
-                                    *((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                        .offset(row as isize),
-                                    pVals[0].data.text_mut_ptr(),
-                                );
-                            }
-                        },
-                        _ => {}
-                    },
-                    FuncOp::SetNull => match (lParse.Nodes[this_node_idx]).ntype {
-                        ValueSort::Long => loop {
-                            let fresh101 = elem;
-                            elem -= 1;
-                            if fresh101 == 0 {
-                                break;
-                            }
-                            if (lParse.Nodes[theParams[1]]).value.data.lng()
-                                == *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                    .offset(elem as isize)
-                            {
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(elem as isize) = 0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                        .offset(elem as isize);
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize);
-                            }
-                        },
-                        ValueSort::Double => loop {
-                            let fresh102 = elem;
-                            elem -= 1;
-                            if fresh102 == 0 {
-                                break;
-                            }
-                            if (lParse.Nodes[theParams[1]]).value.data.dbl()
-                                == *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize)
-                            {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                        .offset(elem as isize);
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) =
-                                    *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize);
-                            }
-                        },
-                        _ => {}
-                    },
-                    FuncOp::Sin => loop {
-                        let fresh103 = elem;
-                        elem -= 1;
-                        if fresh103 == 0 {
-                            break;
-                        }
-                        let fresh104 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh104 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh104 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) =
-                                sin(*((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize));
-                        }
-                    },
-                    FuncOp::Cos => loop {
-                        let fresh105 = elem;
-                        elem -= 1;
-                        if fresh105 == 0 {
-                            break;
-                        }
-                        let fresh106 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh106 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh106 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) =
-                                cos(*((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize));
-                        }
-                    },
-                    FuncOp::Tan => loop {
-                        let fresh107 = elem;
-                        elem -= 1;
-                        if fresh107 == 0 {
-                            break;
-                        }
-                        let fresh108 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh108 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh108 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) =
-                                tan(*((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize));
-                        }
-                    },
-                    FuncOp::Asin => loop {
-                        let fresh109 = elem;
-                        elem -= 1;
-                        if fresh109 == 0 {
-                            break;
-                        }
-                        let fresh110 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh110 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh110 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            if dval < -1.0 || dval > 1.0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = asin(dval);
-                            }
-                        }
-                    },
-                    FuncOp::Acos => loop {
-                        let fresh111 = elem;
-                        elem -= 1;
-                        if fresh111 == 0 {
-                            break;
-                        }
-                        let fresh112 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh112 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh112 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            if dval < -1.0 || dval > 1.0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = acos(dval);
-                            }
-                        }
-                    },
-                    FuncOp::Atan => loop {
-                        let fresh113 = elem;
-                        elem -= 1;
-                        if fresh113 == 0 {
-                            break;
-                        }
-                        let fresh114 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh114 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh114 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = atan(dval);
-                        }
-                    },
-                    FuncOp::Sinh => loop {
-                        let fresh115 = elem;
-                        elem -= 1;
-                        if fresh115 == 0 {
-                            break;
-                        }
-                        let fresh116 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh116 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh116 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = sinh(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize),
-                            );
-                        }
-                    },
-                    FuncOp::Cosh => loop {
-                        let fresh117 = elem;
-                        elem -= 1;
-                        if fresh117 == 0 {
-                            break;
-                        }
-                        let fresh118 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh118 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh118 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = cosh(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize),
-                            );
-                        }
-                    },
-                    FuncOp::Tanh => loop {
-                        let fresh119 = elem;
-                        elem -= 1;
-                        if fresh119 == 0 {
-                            break;
-                        }
-                        let fresh120 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh120 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh120 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = tanh(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize),
-                            );
-                        }
-                    },
-                    FuncOp::Exp => loop {
-                        let fresh121 = elem;
-                        elem -= 1;
-                        if fresh121 == 0 {
-                            break;
-                        }
-                        let fresh122 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh122 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh122 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = exp(dval);
-                        }
-                    },
-                    FuncOp::Log => loop {
-                        let fresh123 = elem;
-                        elem -= 1;
-                        if fresh123 == 0 {
-                            break;
-                        }
-                        let fresh124 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh124 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh124 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            if dval <= 0.0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = log(dval);
-                            }
-                        }
-                    },
-                    FuncOp::Log10 => loop {
-                        let fresh125 = elem;
-                        elem -= 1;
-                        if fresh125 == 0 {
-                            break;
-                        }
-                        let fresh126 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh126 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh126 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            if dval <= 0.0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = log10(dval);
-                            }
-                        }
-                    },
-                    FuncOp::Sqrt => loop {
-                        let fresh127 = elem;
-                        elem -= 1;
-                        if fresh127 == 0 {
-                            break;
-                        }
-                        let fresh128 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh128 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh128 == 0 {
-                            dval = *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                .offset(elem as isize);
-                            if dval < 0.0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = 0.0;
-                                *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize) = 1;
-                            } else {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = sqrt(dval);
-                            }
-                        }
-                    },
-                    FuncOp::Ceil => loop {
-                        let fresh129 = elem;
-                        elem -= 1;
-                        if fresh129 == 0 {
-                            break;
-                        }
-                        let fresh130 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh130 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh130 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = ceil(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize),
-                            );
-                        }
-                    },
-                    FuncOp::Floor => loop {
-                        let fresh131 = elem;
-                        elem -= 1;
-                        if fresh131 == 0 {
-                            break;
-                        }
-                        let fresh132 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh132 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh132 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = floor(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize),
-                            );
-                        }
-                    },
-                    FuncOp::Round => loop {
-                        let fresh133 = elem;
-                        elem -= 1;
-                        if fresh133 == 0 {
-                            break;
-                        }
-                        let fresh134 =
-                            &mut *((lParse.Nodes[this_node_idx]).value.undef).offset(elem as isize);
-                        *fresh134 =
-                            *((lParse.Nodes[theParams[0]]).value.undef).offset(elem as isize);
-                        if *fresh134 == 0 {
-                            *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                .offset(elem as isize) = floor(
-                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                    .offset(elem as isize)
-                                    + 0.5,
-                            );
-                        }
-                    },
-                    FuncOp::Atan2 => loop {
-                        let fresh135 = row;
-                        row -= 1;
-                        if fresh135 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh136 = nelem;
-                            nelem -= 1;
-                            if fresh136 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-
-                            i = 2;
-                            loop {
-                                let fresh137 = i;
-                                i -= 1;
-                                if fresh137 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh138 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh138 = if c_int::from(pNull[0]) != 0 || c_int::from(pNull[1]) != 0
-                            {
-                                1
-                            } else {
-                                0
-                            };
-                            if *fresh138 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) =
-                                    atan2(pVals[0].data.dbl(), pVals[1].data.dbl());
-                            }
-                        }
-                    },
-                    FuncOp::AngSep => loop {
-                        let fresh139 = row;
-                        row -= 1;
-                        if fresh139 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh140 = nelem;
-                            nelem -= 1;
-                            if fresh140 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-                            i = 4 as c_int;
-                            loop {
-                                let fresh141 = i;
-                                i -= 1;
-                                if fresh141 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh142 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh142 = c_int::from(
-                                c_int::from(pNull[0]) != 0
-                                    || c_int::from(pNull[1]) != 0
-                                    || c_int::from(pNull[2]) != 0
-                                    || c_int::from(pNull[3]) != 0,
-                            ) as c_char;
-                            if *fresh142 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(elem as isize) = angsep_calc(
-                                    pVals[0].data.dbl(),
-                                    pVals[1].data.dbl(),
-                                    pVals[2].data.dbl(),
-                                    pVals[3].data.dbl(),
-                                );
-                            }
-                        }
-                    },
-                    FuncOp::Min1 => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                            let mut minVal: c_long = 0;
-                            loop {
-                                let fresh143 = row;
-                                row -= 1;
-                                if fresh143 == 0 {
-                                    break;
-                                }
-                                valInit = 1;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh144 = nelem;
-                                    nelem -= 1;
-                                    if fresh144 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        if valInit != 0 {
-                                            valInit = 0;
-                                            minVal = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(elem as isize);
-                                        } else {
-                                            minVal = if minVal
-                                                < *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(elem as isize)
-                                            {
-                                                minVal
-                                            } else {
-                                                *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                                    .offset(elem as isize)
-                                            };
-                                        }
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = minVal;
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                            let mut minVal_0: c_double = 0.0;
-                            loop {
-                                let fresh145 = row;
-                                row -= 1;
-                                if fresh145 == 0 {
-                                    break;
-                                }
-                                valInit = 1;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh146 = nelem;
-                                    nelem -= 1;
-                                    if fresh146 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        if valInit != 0 {
-                                            valInit = 0;
-                                            minVal_0 = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(elem as isize);
-                                        } else {
-                                            minVal_0 = if minVal_0
-                                                < *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(elem as isize)
-                                            {
-                                                minVal_0
-                                            } else {
-                                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                                    .offset(elem as isize)
-                                            };
-                                        }
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(row as isize) = minVal_0;
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Bits {
-                            let mut minVal_1: c_char = 0;
-                            loop {
-                                let fresh147 = row;
-                                row -= 1;
-                                if fresh147 == 0 {
-                                    break;
-                                }
-                                let mut sptr1_0: *mut c_char =
-                                    *((lParse.Nodes[theParams[0]]).value.data.str_buf())
-                                        .offset(row as isize);
-                                minVal_1 = b'1' as c_char;
-                                while *sptr1_0 != 0 {
-                                    if c_int::from(*sptr1_0) == '0' as i32 {
-                                        minVal_1 = b'0' as c_char;
-                                    }
-                                    sptr1_0 = sptr1_0.offset(1);
-                                }
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(0) = minVal_1;
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(1) = 0;
-                            }
-                        }
-                    }
-                    FuncOp::Min2 => {
-                        if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh148 = row;
-                                row -= 1;
-                                if fresh148 == 0 {
-                                    break;
-                                }
-                                nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                                loop {
-                                    let fresh149 = nelem;
-                                    nelem -= 1;
-                                    if fresh149 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    i = 2;
-                                    loop {
-                                        let fresh150 = i;
-                                        i -= 1;
-                                        if fresh150 == 0 {
-                                            break;
-                                        }
-                                        if vector[i as usize] > 1 {
-                                            pVals[i as usize].data = NodeValue::Long(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(elem as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(elem as isize);
-                                        } else if vector[i as usize] != 0 {
-                                            pVals[i as usize].data = NodeValue::Long(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(row as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(row as isize);
-                                        }
-                                    }
-                                    if c_int::from(pNull[0]) != 0 && c_int::from(pNull[1]) != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = 0;
-                                    } else if pNull[0] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[1].data.lng();
-                                    } else if pNull[1] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[0].data.lng();
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) =
-                                            if pVals[0].data.lng() < pVals[1].data.lng() {
-                                                pVals[0].data.lng()
-                                            } else {
-                                                pVals[1].data.lng()
-                                            };
-                                    }
-                                }
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh151 = row;
-                                row -= 1;
-                                if fresh151 == 0 {
-                                    break;
-                                }
-                                nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                                loop {
-                                    let fresh152 = nelem;
-                                    nelem -= 1;
-                                    if fresh152 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    i = 2;
-                                    loop {
-                                        let fresh153 = i;
-                                        i -= 1;
-                                        if fresh153 == 0 {
-                                            break;
-                                        }
-                                        if vector[i as usize] > 1 {
-                                            pVals[i as usize].data = NodeValue::Double(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(elem as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(elem as isize);
-                                        } else if vector[i as usize] != 0 {
-                                            pVals[i as usize].data = NodeValue::Double(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(row as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(row as isize);
-                                        }
-                                    }
-                                    if c_int::from(pNull[0]) != 0 && c_int::from(pNull[1]) != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = 0.0;
-                                    } else if pNull[0] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[1].data.dbl();
-                                    } else if pNull[1] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[0].data.dbl();
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) =
-                                            if pVals[0].data.dbl() < pVals[1].data.dbl() {
-                                                pVals[0].data.dbl()
-                                            } else {
-                                                pVals[1].data.dbl()
-                                            };
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Max1 => {
-                        elem = row * (lParse.Nodes[theParams[0]]).value.nelem;
-                        if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                            let mut maxVal: c_long = 0;
-                            loop {
-                                let fresh154 = row;
-                                row -= 1;
-                                if fresh154 == 0 {
-                                    break;
-                                }
-                                valInit = 1;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh155 = nelem;
-                                    nelem -= 1;
-                                    if fresh155 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        if valInit != 0 {
-                                            valInit = 0;
-                                            maxVal = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(elem as isize);
-                                        } else {
-                                            maxVal = if maxVal
-                                                > *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(elem as isize)
-                                            {
-                                                maxVal
-                                            } else {
-                                                *((lParse.Nodes[theParams[0]]).value.data.lng_buf())
-                                                    .offset(elem as isize)
-                                            };
-                                        }
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                    .offset(row as isize) = maxVal;
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                            let mut maxVal_0: c_double = 0.0;
-                            loop {
-                                let fresh156 = row;
-                                row -= 1;
-                                if fresh156 == 0 {
-                                    break;
-                                }
-                                valInit = 1;
-                                *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                    1;
-                                nelem = (lParse.Nodes[theParams[0]]).value.nelem;
-                                loop {
-                                    let fresh157 = nelem;
-                                    nelem -= 1;
-                                    if fresh157 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    if *((lParse.Nodes[theParams[0]]).value.undef)
-                                        .offset(elem as isize)
-                                        == 0
-                                    {
-                                        if valInit != 0 {
-                                            valInit = 0;
-                                            maxVal_0 = *((lParse.Nodes[theParams[0]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(elem as isize);
-                                        } else {
-                                            maxVal_0 = if maxVal_0
-                                                > *((lParse.Nodes[theParams[0]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(elem as isize)
-                                            {
-                                                maxVal_0
-                                            } else {
-                                                *((lParse.Nodes[theParams[0]]).value.data.dbl_buf())
-                                                    .offset(elem as isize)
-                                            };
-                                        }
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(row as isize) = 0;
-                                    }
-                                }
-                                *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                    .offset(row as isize) = maxVal_0;
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Bits {
-                            let mut maxVal_1: c_char = 0;
-                            loop {
-                                let fresh158 = row;
-                                row -= 1;
-                                if fresh158 == 0 {
-                                    break;
-                                }
-                                let mut sptr1_1: *mut c_char =
-                                    *((lParse.Nodes[theParams[0]]).value.data.str_buf())
-                                        .offset(row as isize);
-                                maxVal_1 = b'0' as c_char;
-                                while *sptr1_1 != 0 {
-                                    if c_int::from(*sptr1_1) == '1' as i32 {
-                                        maxVal_1 = b'1' as c_char;
-                                    }
-                                    sptr1_1 = sptr1_1.offset(1);
-                                }
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(0) = maxVal_1;
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(1) = 0;
-                            }
-                        }
-                    }
-                    FuncOp::Max2 => {
-                        if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Long {
-                            loop {
-                                let fresh159 = row;
-                                row -= 1;
-                                if fresh159 == 0 {
-                                    break;
-                                }
-                                nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                                loop {
-                                    let fresh160 = nelem;
-                                    nelem -= 1;
-                                    if fresh160 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    i = 2;
-                                    loop {
-                                        let fresh161 = i;
-                                        i -= 1;
-                                        if fresh161 == 0 {
-                                            break;
-                                        }
-                                        if vector[i as usize] > 1 {
-                                            pVals[i as usize].data = NodeValue::Long(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(elem as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(elem as isize);
-                                        } else if vector[i as usize] != 0 {
-                                            pVals[i as usize].data = NodeValue::Long(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .lng_buf())
-                                                .offset(row as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(row as isize);
-                                        }
-                                    }
-                                    if c_int::from(pNull[0]) != 0 && c_int::from(pNull[1]) != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = 0;
-                                    } else if pNull[0] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[1].data.lng();
-                                    } else if pNull[1] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[0].data.lng();
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) =
-                                            if pVals[0].data.lng() > pVals[1].data.lng() {
-                                                pVals[0].data.lng()
-                                            } else {
-                                                pVals[1].data.lng()
-                                            };
-                                    }
-                                }
-                            }
-                        } else if (lParse.Nodes[this_node_idx]).ntype == ValueSort::Double {
-                            loop {
-                                let fresh162 = row;
-                                row -= 1;
-                                if fresh162 == 0 {
-                                    break;
-                                }
-                                nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                                loop {
-                                    let fresh163 = nelem;
-                                    nelem -= 1;
-                                    if fresh163 == 0 {
-                                        break;
-                                    }
-                                    elem -= 1;
-                                    i = 2;
-                                    loop {
-                                        let fresh164 = i;
-                                        i -= 1;
-                                        if fresh164 == 0 {
-                                            break;
-                                        }
-                                        if vector[i as usize] > 1 {
-                                            pVals[i as usize].data = NodeValue::Double(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(elem as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(elem as isize);
-                                        } else if vector[i as usize] != 0 {
-                                            pVals[i as usize].data = NodeValue::Double(
-                                                *((lParse.Nodes[theParams[i as usize]])
-                                                    .value
-                                                    .data
-                                                    .dbl_buf())
-                                                .offset(row as isize),
-                                            );
-                                            pNull[i as usize] = *((lParse.Nodes
-                                                [theParams[i as usize]])
-                                                .value
-                                                .undef)
-                                                .offset(row as isize);
-                                        }
-                                    }
-                                    if c_int::from(pNull[0]) != 0 && c_int::from(pNull[1]) != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 1;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = 0.0;
-                                    } else if pNull[0] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[1].data.dbl();
-                                    } else if pNull[1] != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[0].data.dbl();
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = 0;
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) =
-                                            if pVals[0].data.dbl() > pVals[1].data.dbl() {
-                                                pVals[0].data.dbl()
-                                            } else {
-                                                pVals[1].data.dbl()
-                                            };
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FuncOp::Near => loop {
-                        let fresh165 = row;
-                        row -= 1;
-                        if fresh165 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh166 = nelem;
-                            nelem -= 1;
-                            if fresh166 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-                            i = 3 as c_int;
-                            loop {
-                                let fresh167 = i;
-                                i -= 1;
-                                if fresh167 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh168 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh168 = c_int::from(
-                                c_int::from(pNull[0]) != 0
-                                    || c_int::from(pNull[1]) != 0
-                                    || c_int::from(pNull[2]) != 0,
-                            ) as c_char;
-                            if *fresh168 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                    .offset(elem as isize) = bnear(
-                                    pVals[0].data.dbl(),
-                                    pVals[1].data.dbl(),
-                                    pVals[2].data.dbl(),
-                                );
-                            }
-                        }
-                    },
-                    FuncOp::Circle => loop {
-                        let fresh169 = row;
-                        row -= 1;
-                        if fresh169 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh170 = nelem;
-                            nelem -= 1;
-                            if fresh170 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-                            i = 5 as c_int;
-                            loop {
-                                let fresh171 = i;
-                                i -= 1;
-                                if fresh171 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh172 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh172 = c_int::from(
-                                c_int::from(pNull[0]) != 0
-                                    || c_int::from(pNull[1]) != 0
-                                    || c_int::from(pNull[2]) != 0
-                                    || c_int::from(pNull[3]) != 0
-                                    || c_int::from(pNull[4]) != 0,
-                            ) as c_char;
-                            if *fresh172 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                    .offset(elem as isize) = circle(
-                                    pVals[0].data.dbl(),
-                                    pVals[1].data.dbl(),
-                                    pVals[2].data.dbl(),
-                                    pVals[3].data.dbl(),
-                                    pVals[4].data.dbl(),
-                                );
-                            }
-                        }
-                    },
-                    FuncOp::Box => loop {
-                        let fresh173 = row;
-                        row -= 1;
-                        if fresh173 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh174 = nelem;
-                            nelem -= 1;
-                            if fresh174 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-                            i = 7 as c_int;
-                            loop {
-                                let fresh175 = i;
-                                i -= 1;
-                                if fresh175 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh176 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh176 = c_int::from(
-                                c_int::from(pNull[0]) != 0
-                                    || c_int::from(pNull[1]) != 0
-                                    || c_int::from(pNull[2]) != 0
-                                    || c_int::from(pNull[3]) != 0
-                                    || c_int::from(pNull[4]) != 0
-                                    || c_int::from(pNull[5]) != 0
-                                    || c_int::from(pNull[6]) != 0,
-                            ) as c_char;
-                            if *fresh176 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                    .offset(elem as isize) = saobox(
-                                    pVals[0].data.dbl(),
-                                    pVals[1].data.dbl(),
-                                    pVals[2].data.dbl(),
-                                    pVals[3].data.dbl(),
-                                    pVals[4].data.dbl(),
-                                    pVals[5].data.dbl(),
-                                    pVals[6].data.dbl(),
-                                );
-                            }
-                        }
-                    },
-                    FuncOp::Ellipse => loop {
-                        let fresh177 = row;
-                        row -= 1;
-                        if fresh177 == 0 {
-                            break;
-                        }
-                        nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                        loop {
-                            let fresh178 = nelem;
-                            nelem -= 1;
-                            if fresh178 == 0 {
-                                break;
-                            }
-                            elem -= 1;
-                            i = 7 as c_int;
-                            loop {
-                                let fresh179 = i;
-                                i -= 1;
-                                if fresh179 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] > 1 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(elem as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(elem as isize);
-                                } else if vector[i as usize] != 0 {
-                                    pVals[i as usize].data = NodeValue::Double(
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .dbl_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh180 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(elem as isize);
-                            *fresh180 = c_int::from(
-                                c_int::from(pNull[0]) != 0
-                                    || c_int::from(pNull[1]) != 0
-                                    || c_int::from(pNull[2]) != 0
-                                    || c_int::from(pNull[3]) != 0
-                                    || c_int::from(pNull[4]) != 0
-                                    || c_int::from(pNull[5]) != 0
-                                    || c_int::from(pNull[6]) != 0,
-                            ) as c_char;
-                            if *fresh180 == 0 {
-                                *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                    .offset(elem as isize) = ellipse(
-                                    pVals[0].data.dbl(),
-                                    pVals[1].data.dbl(),
-                                    pVals[2].data.dbl(),
-                                    pVals[3].data.dbl(),
-                                    pVals[4].data.dbl(),
-                                    pVals[5].data.dbl(),
-                                    pVals[6].data.dbl(),
-                                );
-                            }
-                        }
-                    },
-                    FuncOp::IfThenElse => match (lParse.Nodes[this_node_idx]).ntype {
-                        ValueSort::Boolean => loop {
-                            let fresh181 = row;
-                            row -= 1;
-                            if fresh181 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh182 = nelem;
-                                nelem -= 1;
-                                if fresh182 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-                                if vector[2] > 1 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(elem as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(elem as isize);
-                                } else if vector[2] != 0 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(row as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(row as isize);
-                                }
-                                i = 2;
-                                loop {
-                                    let fresh183 = i;
-                                    i -= 1;
-                                    if fresh183 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pVals[i as usize].data = NodeValue::Logical(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .log_buf())
-                                            .offset(elem as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                    } else if vector[i as usize] != 0 {
-                                        pVals[i as usize].data = NodeValue::Logical(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .log_buf())
-                                            .offset(row as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                    }
-                                }
-                                let fresh184 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize);
-                                *fresh184 = pNull[2];
-                                if *fresh184 == 0 {
-                                    if pVals[2].data.log() != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                            .offset(elem as isize) = pVals[0].data.log();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[0];
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.data.log_buf())
-                                            .offset(elem as isize) = pVals[1].data.log();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[1];
-                                    }
-                                }
-                            }
-                        },
-                        ValueSort::Long => loop {
-                            let fresh185 = row;
-                            row -= 1;
-                            if fresh185 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh186 = nelem;
-                                nelem -= 1;
-                                if fresh186 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-                                if vector[2] > 1 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(elem as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(elem as isize);
-                                } else if vector[2] != 0 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(row as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(row as isize);
-                                }
-                                i = 2;
-                                loop {
-                                    let fresh187 = i;
-                                    i -= 1;
-                                    if fresh187 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pVals[i as usize].data = NodeValue::Long(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(elem as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                    } else if vector[i as usize] != 0 {
-                                        pVals[i as usize].data = NodeValue::Long(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .lng_buf())
-                                            .offset(row as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                    }
-                                }
-                                let fresh188 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize);
-                                *fresh188 = pNull[2];
-                                if *fresh188 == 0 {
-                                    if pVals[2].data.log() != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[0].data.lng();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[0];
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                            .offset(elem as isize) = pVals[1].data.lng();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[1];
-                                    }
-                                }
-                            }
-                        },
-                        ValueSort::Double => loop {
-                            let fresh189 = row;
-                            row -= 1;
-                            if fresh189 == 0 {
-                                break;
-                            }
-                            nelem = (lParse.Nodes[this_node_idx]).value.nelem;
-                            loop {
-                                let fresh190 = nelem;
-                                nelem -= 1;
-                                if fresh190 == 0 {
-                                    break;
-                                }
-                                elem -= 1;
-                                if vector[2] > 1 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(elem as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(elem as isize);
-                                } else if vector[2] != 0 {
-                                    pVals[2].data = NodeValue::Logical(
-                                        *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                            .offset(row as isize),
-                                    );
-                                    pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                        .offset(row as isize);
-                                }
-                                i = 2;
-                                loop {
-                                    let fresh191 = i;
-                                    i -= 1;
-                                    if fresh191 == 0 {
-                                        break;
-                                    }
-                                    if vector[i as usize] > 1 {
-                                        pVals[i as usize].data = NodeValue::Double(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(elem as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(elem as isize);
-                                    } else if vector[i as usize] != 0 {
-                                        pVals[i as usize].data = NodeValue::Double(
-                                            *((lParse.Nodes[theParams[i as usize]])
-                                                .value
-                                                .data
-                                                .dbl_buf())
-                                            .offset(row as isize),
-                                        );
-                                        pNull[i as usize] =
-                                            *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                                .offset(row as isize);
-                                    }
-                                }
-                                let fresh192 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                    .offset(elem as isize);
-                                *fresh192 = pNull[2];
-                                if *fresh192 == 0 {
-                                    if pVals[2].data.log() != 0 {
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[0].data.dbl();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[0];
-                                    } else {
-                                        *((lParse.Nodes[this_node_idx]).value.data.dbl_buf())
-                                            .offset(elem as isize) = pVals[1].data.dbl();
-                                        *((lParse.Nodes[this_node_idx]).value.undef)
-                                            .offset(elem as isize) = pNull[1];
-                                    }
-                                }
-                            }
-                        },
-                        ValueSort::String => loop {
-                            let fresh193 = row;
-                            row -= 1;
-                            if fresh193 == 0 {
-                                break;
-                            }
-                            if vector[2] != 0 {
-                                pVals[2].data = NodeValue::Logical(
-                                    *((lParse.Nodes[theParams[2]]).value.data.log_buf())
-                                        .offset(row as isize),
-                                );
-                                pNull[2] = *((lParse.Nodes[theParams[2]]).value.undef)
-                                    .offset(row as isize);
-                            }
-                            i = 2;
-                            loop {
-                                let fresh194 = i;
-                                i -= 1;
-                                if fresh194 == 0 {
-                                    break;
-                                }
-                                if vector[i as usize] != 0 {
-                                    strcpy(
-                                        pVals[i as usize].data.text_mut_ptr(),
-                                        *((lParse.Nodes[theParams[i as usize]])
-                                            .value
-                                            .data
-                                            .str_buf())
-                                        .offset(row as isize),
-                                    );
-                                    pNull[i as usize] =
-                                        *((lParse.Nodes[theParams[i as usize]]).value.undef)
-                                            .offset(row as isize);
-                                }
-                            }
-                            let fresh195 = &mut *((lParse.Nodes[this_node_idx]).value.undef)
-                                .offset(row as isize);
-                            *fresh195 = pNull[2];
-                            if *fresh195 == 0 {
-                                if pVals[2].data.log() != 0 {
-                                    strcpy(
-                                        *((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                            .offset(row as isize),
-                                        pVals[0].data.text_mut_ptr(),
-                                    );
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = pNull[0];
-                                } else {
-                                    strcpy(
-                                        *((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                            .offset(row as isize),
-                                        pVals[1].data.text_mut_ptr(),
-                                    );
-                                    *((lParse.Nodes[this_node_idx]).value.undef)
-                                        .offset(row as isize) = pNull[1];
-                                }
-                            } else {
-                                *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                    .offset(row as isize))
-                                .offset(0) = 0;
-                            }
-                        },
-                        _ => {}
-                    },
-                    FuncOp::StrMid => {
-                        let strconst: c_int =
-                            c_int::from((lParse.Nodes[theParams[0]]).operation == Operation::Const);
-                        let posconst: c_int =
-                            c_int::from((lParse.Nodes[theParams[1]]).operation == Operation::Const);
-                        let lenconst: c_int =
-                            c_int::from((lParse.Nodes[theParams[2]]).operation == Operation::Const);
-                        let dest_len: c_int = (lParse.Nodes[this_node_idx]).value.nelem as c_int;
-                        let mut src_len: c_int = (lParse.Nodes[theParams[0]]).value.nelem as c_int;
-                        loop {
-                            let fresh196 = row;
-                            row -= 1;
-                            if fresh196 == 0 {
-                                break;
-                            }
-                            let mut pos: c_int = 0;
-                            let mut len: c_int = 0;
-                            let mut str: *mut c_char = ptr::null_mut();
-                            let mut undef: c_int = 0;
-                            if posconst != 0 {
-                                pos = (lParse.Nodes[theParams[1]]).value.data.lng() as c_int;
-                            } else {
-                                pos = *((lParse.Nodes[theParams[1]]).value.data.lng_buf())
-                                    .offset(row as isize)
-                                    as c_int;
-                                if *((lParse.Nodes[theParams[1]]).value.undef).offset(row as isize)
-                                    != 0
-                                {
-                                    undef = 1;
-                                }
-                            }
-                            if strconst != 0 {
-                                str = (lParse.Nodes[theParams[0]]).value.data.text_mut_ptr();
-                                if src_len == 0 {
-                                    src_len = strlen(str) as c_int;
-                                }
-                            } else {
-                                str = *((lParse.Nodes[theParams[0]]).value.data.str_buf())
-                                    .offset(row as isize);
-                                if *((lParse.Nodes[theParams[0]]).value.undef).offset(row as isize)
-                                    != 0
-                                {
-                                    undef = 1;
-                                }
-                            }
-                            if lenconst != 0 {
-                                len = dest_len;
-                            } else {
-                                len = *((lParse.Nodes[theParams[2]]).value.data.lng_buf())
-                                    .offset(row as isize)
-                                    as c_int;
-                                if *((lParse.Nodes[theParams[2]]).value.undef).offset(row as isize)
-                                    != 0
-                                {
-                                    undef = 1;
-                                }
-                            }
-                            *(*((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                .offset(row as isize))
-                            .offset(0) = 0;
-                            if pos == 0 {
-                                undef = 1;
-                            }
-                            if undef == 0
-                                && cstrmid(
-                                    lParse,
-                                    *((lParse.Nodes[this_node_idx]).value.data.str_buf())
-                                        .offset(row as isize),
-                                    len,
-                                    str,
-                                    src_len,
-                                    pos,
-                                ) < 0
-                            {
-                                break;
-                            }
-                            *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                undef as c_char;
-                        }
-                    }
-                    FuncOp::StrPos => {
-                        let const1: c_int =
-                            c_int::from((lParse.Nodes[theParams[0]]).operation == Operation::Const);
-                        let const2: c_int =
-                            c_int::from((lParse.Nodes[theParams[1]]).operation == Operation::Const);
-                        loop {
-                            let fresh197 = row;
-                            row -= 1;
-                            if fresh197 == 0 {
-                                break;
-                            }
-                            let mut str1: *mut c_char = ptr::null_mut();
-                            let mut str2: *mut c_char = ptr::null_mut();
-                            let mut undef_0: c_int = 0;
-                            if const1 != 0 {
-                                str1 = (lParse.Nodes[theParams[0]]).value.data.text_mut_ptr();
-                            } else {
-                                str1 = *((lParse.Nodes[theParams[0]]).value.data.str_buf())
-                                    .offset(row as isize);
-                                if *((lParse.Nodes[theParams[0]]).value.undef).offset(row as isize)
-                                    != 0
-                                {
-                                    undef_0 = 1;
-                                }
-                            }
-                            if const2 != 0 {
-                                str2 = (lParse.Nodes[theParams[1]]).value.data.text_mut_ptr();
-                            } else {
-                                str2 = *((lParse.Nodes[theParams[1]]).value.data.str_buf())
-                                    .offset(row as isize);
-                                if *((lParse.Nodes[theParams[1]]).value.undef).offset(row as isize)
-                                    != 0
-                                {
-                                    undef_0 = 1;
-                                }
-                            }
-                            *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                .offset(row as isize) = 0;
-                            if undef_0 == 0 {
-                                let res_0: *mut c_char = strstr(str1, str2);
-                                if res_0.is_null() {
-                                    undef_0 = 1;
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(row as isize) = 0;
-                                } else {
-                                    *((lParse.Nodes[this_node_idx]).value.data.lng_buf())
-                                        .offset(row as isize) =
-                                        res_0.offset_from(str1) as c_long + 1;
-                                }
-                            }
-                            *((lParse.Nodes[this_node_idx]).value.undef).offset(row as isize) =
-                                undef_0 as c_char;
-                        }
-                    }
+                    FuncOp::Row => func::row_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Null => func::null_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::AxisElem => func::axis_elem_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::ElemNum => func::elem_num_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Rnd => func::rnd_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::GasRnd => func::gas_rnd_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::PoiRnd => func::poi_rnd_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Sum => func::sum_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Average => func::average_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Stddev => func::stddev_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Median => func::median_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Abs => func::abs_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::NonNull => func::non_null_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::IsNull => func::is_null_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::DefNull => func::def_null_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::SetNull => func::set_null_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Sin => func::sin_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Cos => func::cos_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Tan => func::tan_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Asin => func::asin_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Acos => func::acos_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Atan => func::atan_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Sinh => func::sinh_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Cosh => func::cosh_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Tanh => func::tanh_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Exp => func::exp_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Log => func::log_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Log10 => func::log10_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Sqrt => func::sqrt_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Ceil => func::ceil_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Floor => func::floor_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Round => func::round_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Atan2 => func::atan2_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::AngSep => func::ang_sep_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Min1 => func::min1_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Min2 => func::min2_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Max1 => func::max1_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Max2 => func::max2_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Near => func::near_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Circle => func::circle_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Box => func::box_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::Ellipse => func::ellipse_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::IfThenElse => func::if_then_else_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::StrMid => func::str_mid_rows(lParse, this_node_idx, &mut args),
+                    FuncOp::StrPos => func::str_pos_rows(lParse, this_node_idx, &mut args),
                     _ => {}
                 }
             }
@@ -7160,11 +4038,11 @@ fn Do_Func(lParse: &mut ParseData, this_node_idx: usize) {
             if fresh198 == 0 {
                 break;
             }
-            if (lParse.Nodes[theParams[i as usize]])
+            if (lParse.Nodes[args.params[i as usize]])
                 .operation
                 .is_computed()
             {
-                free((lParse.Nodes[theParams[i as usize]]).value.data.raw());
+                free((lParse.Nodes[args.params[i as usize]]).value.data.raw());
             }
         }
     }
